Solve task 16

+++ 0x02-functions_nested_loops/104-fibonacci.c
@@ -21,7 +21,7 @@ int main(void)
 	for (i = 3; i <= 98; i++)
 	{
 		long int c1 = ((long)a1 + b1) / 1000000000;
-		long int c2 = (a2 + b2) % 1000000000;
+		long int c2 = ((long)a2 + b2) % 1000000000;
 
 		if (c1)
 		{
